Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch test_bot
Changes to be committed:
	modified:   .gitignore
	modified:   __pycache__/app.cpython-311.pyc
	modified:   __pycache__/config.cpython-311.pyc
	modified:   __pycache__/keyboards.cpython-311.pyc
	modified:   app.py
	modified:   bot.py
	deleted:    clear_db/delete_question.py
	deleted:    clear_db/status_replacement.py
	new file:   delete_question.py
	modified:   keyboards.py
	modified:   main.db
	new file:   sending.service
	new file:   sending_answer.py
	new file:   test_bot.service
	modified:   utils/__pycache__/statesQuestion.cpython-311.pyc
	modified:   utils/statesQuestion.py
